move publish test and improve mockLogs
