+++ src/__fixtures__/mockLogs.ts
@@ -1,47 +1,51 @@
+/** Methods that will be mocked. More could be added later if needed. */
+export type MockLogMethod = 'log' | 'warn' | 'error';
+const mockedMethods: MockLogMethod[] = ['log', 'warn', 'error'];
+
 export type MockLogs = {
-  /** Mocked methods (to access calls etc). More could be added later if needed. */
-  mocks: { [k in 'log' | 'warn' | 'error']: jest.SpyInstance<void, Parameters<typeof console.log>> };
+  /** Mocked methods (to access calls etc) */
+  mocks: { [k in MockLogMethod]: jest.SpyInstance<void, Parameters<typeof console.log>> };
   /** Actual console methods */
   realConsole: typeof console;
-  /** Re-init the mocks (you only need to manually call this after resetting or restoring mocks) */
-  init: (alsoLog?: boolean) => void;
-  /** Clear mock results */
-  clear: () => void;
+  /** Re-init the mocks for the current test (to use different options for this test) */
+  init: (alsoLog?: boolean | MockLogMethod[]) => void;
   /** Restore mock implementations */
   restore: () => void;
 };
 
 /**
- * Initialize console log mocks, which will be cleared (but not reset) after each test.
- * This should be called **outside** of any lifecycle hooks or tests because it calls `beforeAll`,
- * `afterEach`, and `afterAll` for setup and teardown.
- * @param alsoLog Whether to also log to the real console. Can be controlled by VERBOSE env var.
+ * Initialize console log mocks, which will be reset after each test. This should be called **outside**
+ * of any lifecycle hooks or tests because it calls `beforeEach` and `afterEach` for setup and teardown.
+ * @param alsoLog Whether to also log to the real console (or subset of methods to log to the console).
+ * All logging can be enabled by setting the VERBOSE env var.
  */
-export function initMockLogs(alsoLog: boolean = !!process.env.VERBOSE): MockLogs {
-  const mock = (method: 'log' | 'warn' | 'error', shouldLog: boolean) => {
-    logs.mocks[method] = jest
-      .spyOn(console, method)
-      .mockImplementation((...args) => (shouldLog ? realConsole[method](...args) : undefined));
-  };
-
-  const realConsole = { ...console };
+export function initMockLogs(alsoLog?: boolean | MockLogMethod[]): MockLogs {
   const logs: MockLogs = {
     mocks: {} as MockLogs['mocks'],
-    realConsole,
-    init: (shouldLog = alsoLog) => (['log', 'warn', 'error'] as const).forEach(method => mock(method, shouldLog)),
-    clear: () => Object.values(logs.mocks).forEach(mock => mock.mockClear()),
-    restore: () => Object.values(logs.mocks).forEach(mock => mock.mockRestore()),
+    realConsole: { ...console },
+    init: (shouldLog = alsoLog) => {
+      logs.restore(); // clear any previous mocks
+      mockedMethods.forEach(method => {
+        const shouldLogMethod = typeof shouldLog === 'boolean' ? shouldLog : shouldLog?.includes(method);
+        logs.mocks[method] = jest
+          .spyOn(console, method)
+          .mockImplementation((...args) =>
+            shouldLogMethod || process.env.VERBOSE ? logs.realConsole[method](...args) : undefined
+          );
+      });
+    },
+    restore: () =>
+      Object.entries(logs.mocks).forEach(([name, mock]) => {
+        mock.mockRestore();
+        delete (logs.mocks as any)[name];
+      }),
   };
 
-  beforeAll(() => {
+  beforeEach(() => {
     logs.init(alsoLog);
   });
 
   afterEach(() => {
-    logs.clear();
-  });
-
-  afterAll(() => {
     logs.restore();
   });
